requirements/default.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/default.txt requirements.in/default.in
#
amqp==2.4.2
asn1crypto==0.24.0        # via cryptography
babel==2.6.0              # via flower
bcrypt==3.1.6             # via paramiko
billiard==3.5.0.5         # via celery
boto3==1.9.120
boto==2.49.0
botocore==1.12.120        # via boto3, s3transfer
cachetools==3.1.0         # via google-auth
celery==4.2.2
cerberus==1.2
certifi==2019.3.9         # via requests
cffi==1.12.2              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
cryptography==2.6.1       # via fabric, paramiko
django-anymail==6.0
django-appconf==1.0.3     # via django-imagekit
django-cors-headers==2.5.2
django-imagekit==4.0.2
django-storages==1.7.1
django==1.11.20
djangorestframework==3.9.2
docutils==0.14            # via botocore
drf-nested-routers==0.91
drf-writable-nested==0.5.1
easy-thumbnails==2.6
elasticsearch-dsl==6.3.1
elasticsearch==6.3.1
fabric==2.4.0
flower==0.9.3
google-api-core==1.8.2    # via google-cloud-core, google-cloud-storage
google-auth==1.6.3        # via google-api-core
google-cloud-core==0.29.1  # via google-cloud-storage
google-cloud-storage==1.14.0
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos==1.5.9  # via google-api-core
idna==2.8                 # via requests
invoke==1.2.0             # via fabric
jmespath==0.9.4           # via boto3, botocore
kombu==4.3.0              # via celery
paramiko==2.4.2           # via fabric
pbr==5.1.3                # via stevedore
pilkit==2.0               # via django-imagekit
pillow==5.4.1             # via easy-thumbnails
pip-tools==3.5.0
protobuf==3.7.0           # via google-api-core, googleapis-common-protos
psycopg2==2.7.7
pyasn1-modules==0.2.4     # via google-auth
pyasn1==0.4.5             # via paramiko, pyasn1-modules, rsa
pycparser==2.19           # via cffi
pyjwt==1.7.1              # via twilio
pynacl==1.3.0             # via paramiko
pysocks==1.6.8            # via twilio
python-dateutil==2.8.0    # via botocore, elasticsearch-dsl
pytz==2018.9              # via babel, celery, django, flower, google-api-core, twilio
pyyaml==5.1
requests==2.21.0
rsa==4.0                  # via google-auth
s3transfer==0.2.0         # via boto3
simplegeneric==0.8.1
six==1.12.0               # via bcrypt, cryptography, django-anymail, django-appconf, django-imagekit, elasticsearch-dsl, google-api-core, google-auth, google-resumable-media, pip-tools, protobuf, pynacl, python-dateutil, stevedore, twilio
stevedore==1.30.1
tornado==5.1.1            # via flower
twilio==6.25.2
urllib3==1.24.1           # via botocore, elasticsearch, requests
vine==1.3.0               # via amqp
